feat: CalendarHelper, update dependencies, rename, and other improvements
This commit introduces the `CalendarHelper` object with functions to interact with the Android Calendar Provider, including adding, reading, updating, and deleting events, as well as querying calendars. It also updates dependencies, including Android Gradle Plugin to 8.8.0, Kotlin Gradle Plugin to 2.1.0, and several AndroidX libraries. Additionally, it renames `SaveableActivity` to `SavableActivity`, adds a dialog with a timer, and includes biometric authentication functionality in the sample app. Other changes involve drawable to bitmap conversion and code cleanup.